requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
supabase>=2.3.1
flask>=2.0.0
